Refactor main Used solid priciples
