+++ main.rb
@@ -1,24 +1,14 @@
 require './app'
-
-def menu_display
-  puts " Please choose an option by entering a number:
-  1 -List all books
-  2 -List all people
-  3 -Create a person
-  4 -Create a book
-  5 -Create a rental
-  6 -List all rentals for a given person id
-  7 -Exit"
-end
+require './utilities'
 
 collectable = { books: [], person: [] }
-
 def main(store)
   app = App.new
+  util = Util.new
   option = nil
 
   while option != '7'
-    menu_display
+    util.menu_display
     option = gets.chomp
     case option
     when '1'
